+++ packages/angular/CHANGELOG.md
@@ -12,7 +12,7 @@
 
 * The following workspace dependencies were updated
   * peerDependencies
-    * @cdssnc/gcds-components bumped from ^0.29.0 to ^0.30.0
+    * @cdssnc/gcds-components bumped from ^0.29.1 to ^0.30.0
 
 ## [0.29.1](https://github.com/cds-snc/gcds-components/compare/gcds-components-angular-v0.29.0...gcds-components-angular-v0.29.1) (2024-12-11)