+++ packages/react/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.30.0](https://github.com/cds-snc/gcds-components/compare/gcds-components-react-v0.29.0...gcds-components-react-v0.30.0) (2024-12-11)
+## [0.30.0](https://github.com/cds-snc/gcds-components/compare/gcds-components-react-v0.29.1...gcds-components-react-v0.30.0) (2024-12-17)
 
 
 ### Patch
@@ -12,7 +12,7 @@
 
 * The following workspace dependencies were updated
   * dependencies
-    * @cdssnc/gcds-components bumped from ^0.29.0 to ^0.30.0
+    * @cdssnc/gcds-components bumped from ^0.29.1 to ^0.30.0
 
 ## [0.29.1](https://github.com/cds-snc/gcds-components/compare/gcds-components-react-v0.29.0...gcds-components-react-v0.29.1) (2024-12-11)